rrdcached: Fix permissions of the default socket.
If neither the "-l" nor the "-P" options have been given, a UNIX domain
socket at a default location is created. However, due to a missing
initialization in this case, the "permissions" field will be left at
zero (no command allowed).

This patch fixes this problem by allowing all commands on the default
socket if no "-P" option is given.